+++ src/ui/components/common/actionables/actions/ChevronDownActionIcon.tsx
@@ -1,6 +1,6 @@
 import { ActionIcon } from "@mantine/core";
 import { IconChevronDown } from "@tabler/icons-react";
-import { useAppearance } from "../../../../services/AppearanceProvider";
+import { useAppearance } from "../../../../providers/AppearanceProvider";
 
 interface ChevronDownActionIconProps {
     tooltip?: string;
