+++ src/ui/components/common/actionables/actions/CopyActionIcon.tsx
@@ -1,6 +1,6 @@
 import { ActionIcon } from "@mantine/core";
 import { IconCopy } from "@tabler/icons-react";
-import { useAppearance } from "../../../../services/AppearanceProvider";
+import { useAppearance } from "../../../../providers/AppearanceProvider";
 
 interface CopyActionIconProps {
     tooltip?: string;
