+++ src/ui/components/scene-manager/views-sidebar/Views.tsx
@@ -1,10 +1,10 @@
 import { useEffect, useMemo, useState, useRef } from "react";
-import { useRegime } from "../../../services/RegimeProvider";
+import { useRegime } from "../../../providers/RegimeProvider";
 import { ViewCard } from "../../view-card/ViewCard";
 import type { Subscription } from "rxjs";
 import { InactiveViewCard } from "../../view-card/InactiveViewCard";
 import { CreateViewCard } from "../../view-card/CreateViewCard";
-import { useDialogue } from "../../../services/DialogueProvider";
+import { useDialogue } from "../../../providers/DialogueProvider";
 import {
     ViewOptionsDialogueContent,
     type ViewOptionsDialogueContentReturnType,
@@ -12,7 +12,7 @@ import {
 import { UiLocalStorageService } from "../../../services/UiLocalStorageService";
 import { pushErrorNotification } from "../../../services/NotificationService";
 import { loggerUi } from "../../../utils/loggerUi";
-import { useManagedAssets } from "../../../services/ManagedAssetsProvider";
+import { useManagedAssets } from "../../../providers/ManagedAssetsProvider";
 import { DeleteViewDialogueContent } from "./DeleteViewDialogueContent";
 import {
     addEmptySnapshotToTree,
