+++ src/ui/main.tsx
@@ -1,13 +1,13 @@
 import { createMemoryRouter, RouterProvider } from "react-router-dom";
 import { createRoot } from "react-dom/client";
 import routes from "./router/routes.tsx";
-import { UserSettingsProvider } from "./services/UserSettingsProvider.tsx";
-import { AppearanceProvider } from "./services/AppearanceProvider.tsx";
+import { UserSettingsProvider } from "./providers/UserSettingsProvider.tsx";
+import { AppearanceProvider } from "./providers/AppearanceProvider.tsx";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { RegimeProvider } from "./services/RegimeProvider.tsx";
-import { DialogueProvider } from "./services/DialogueProvider.tsx";
-import { ManagedAssetsProvider } from "./services/ManagedAssetsProvider.tsx";
-import { ProcessingProvider } from "./services/ProcessingProvider.tsx";
+import { RegimeProvider } from "./providers/RegimeProvider.tsx";
+import { DialogueProvider } from "./providers/DialogueProvider.tsx";
+import { ManagedAssetsProvider } from "./providers/ManagedAssetsProvider.tsx";
+import { ProcessingProvider } from "./providers/ProcessingProvider.tsx";
 
 import "@mantine/core/styles.css";
 import "@mantine/notifications/styles.css";
