+++ src/ui/providers/MenuProvider.tsx
@@ -36,7 +36,7 @@ import { useNavigate, type NavigateFunction } from "react-router-dom";
 import { useFileManagement } from "../hooks/useFileManagement";
 import { AboutDialogueContent } from "../components/common/dialogue/AboutDialogueContent";
 import { useDialogue, type DialogueProps } from "./DialogueProvider";
-import { pushErrorNotification } from "./NotificationService";
+import { pushErrorNotification } from "../services/NotificationService";
 
 /**
  * The action can be either:
